[timeseries] Decouple query dispatcher (#17474)
